pyproject.toml
[tool.poetry]
name = "llm-app-agent"
version = "0.1.0"
description = ""
authors = ["sjchin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyterlab = "^4.2.5"
python-dotenv = "^1.0.1"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langgraph = "^0.2.16"
langchain-community = "^0.2.16"
uvicorn = "^0.30.6"
fastapi = "^0.112.2"
langserve = "^0.2.3"
sse-starlette = "^2.1.3"
streamlit = "^1.38.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
